INS-2208: fixed response helper tests
