+++ conveyor/adapter/response_test.go
@@ -23,9 +23,7 @@ import (
 	"time"
 
 	"github.com/insolar/insolar/conveyor/adapter/adapterid"
-	"github.com/insolar/insolar/conveyor/fsm"
 	"github.com/insolar/insolar/insolar"
-	"github.com/insolar/insolar/messagebus"
 	"github.com/insolar/insolar/testutils"
 	"github.com/stretchr/testify/require"
 )
@@ -140,47 +138,4 @@ func TestResponseSendAdapter_Parallel(t *testing.T) {
 
 	adapter.StopProcessing()
 
-}
-
-type mockReply struct {
-	data string
-}
-
-func (mr *mockReply) Type() insolar.ReplyType {
-	return 0
-}
-
-func TestSendResponseHelper(t *testing.T) {
-	f := messagebus.NewFuture()
-	event := insolar.ConveyorPendingMessage{Future: f}
-	testReply := &mockReply{data: "Put-in"}
-
-	slotElementHelperMock := fsm.NewSlotElementHelperMock(t)
-	slotElementHelperMock.GetInputEventFunc = func() (r interface{}) {
-		return event
-	}
-	slotElementHelperMock.SendTaskFunc = func(p adapterid.ID, response interface{}, p2 uint32) (r error) {
-		f := response.(SendResponseTask).Future
-		f.SetResult(testReply)
-		return nil
-	}
-
-	adapterCatalog := NewHelperCatalog()
-	err := adapterCatalog.SendResponseHelper.SendResponse(slotElementHelperMock, testReply, 42)
-	require.NoError(t, err)
-
-	gotReply, err := f.GetResult(time.Second)
-	require.NoError(t, err)
-	require.Equal(t, testReply, gotReply)
-}
-
-func TestSendResponseHelper_BadInput(t *testing.T) {
-	slotElementHelperMock := fsm.NewSlotElementHelperMock(t)
-	slotElementHelperMock.GetInputEventFunc = func() (r interface{}) {
-		return 33
-	}
-	adapterCatalog := NewHelperCatalog()
-	err := adapterCatalog.SendResponseHelper.SendResponse(slotElementHelperMock, &mockReply{}, 44)
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "Input event is not insolar.ConveyorPendingMessage")
 }
